requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
camelot-py
pdfplumber
tabula-py
pydantic
python-multipart
